Apply TT4 rounded borders also to other product grids
The previous commit mentions that TT4 border-radius is not applied
on blocks over which the users have no control on purpose.

The intention was to not cause any frustration to the user. However,
after testing, I thought it'd be better to actually do the opposite:
it'd be frustrating for the user to be unable to apply the radius that
is the default everywhere else and, after all, the user has chosen
a theme for a certain aesthetic.